+++ src/main/java/com/web/Mlog/repository/PostRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface PostRepository extends JpaRepository<Post, Long> {
+public interface PostRepository extends JpaRepository<Post, Integer> {
     public List<Post> findAllByVisibleIsTrue();
-    public Optional<Post> findByPostIdAndVisibleIsTrue(long id);
+    public Optional<Post> findByPostIdAndVisibleIsTrue(int id);
 }
